+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+- Fix EPG entries for Duration schedule items that play multiple items 
 
 ## [0.2.1-alpha] - 2021-10-24
 ### Fixed