+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+
+## [0.2.2-alpha] - 2021-10-30
 ### Fixed
 - Fix EPG entries for Duration schedule items that play multiple items
 - Fix EPG entries for Multiple schedule items that play more than one item